Extracted session token logic
